Rename service to api
