+++ apps/frontend/src/api/auth/authService.ts
@@ -1,5 +1,5 @@
 import { API_ENDPOINTS } from "@/constants/endpoints";
-import api from "./api";
+import api from "../api";
 
 export const loginService = async (email: string, password: string) => {
   try {
